Cache the 8B qualitative rows and preserve the score dtype
Two residual mismatches from the clean-clone test: the 8B qualitative arms were selected by probing
for generation files that are not shipped, so the table came out empty, and the cached SAFT scores
were read back as float64 where they are computed as float32, which moved the kernel density in the
last bits. Both fixed; all figures and tables now rebuild identically from a clone.

+++ paper/scripts/make_figures.py
@@ -197,7 +197,7 @@ def saft_kde_figure():
         scores = [_score(Z["layer15"]), _score(Z["layer32"]), _score(Z05)]
         json.dump([s.tolist() for s in scores], open(cache, "w"))
     else:
-        scores = [np.array(s) for s in json.load(open(cache))]
+        scores = [np.array(s, dtype=np.float32) for s in json.load(open(cache))]   # as computed
     panels = [("8B, layer 15", scores[0], unsafe, req), ("8B, layer 32", scores[1], unsafe, req), ("0.5B, layer 12", scores[2], unsafe05, req05)]
     fig, ax = plt.subplots(1, 3, figsize=(6.4, 2.1), sharey=False)
     for a, (title, sc, uns, rq) in zip(ax, panels):
